Refactor setting elements
